Refactor member bindings
