+++ src/QueryMutator.Core/MemberBindings/IgnoreMemberBinding.cs
@@ -4,7 +4,7 @@ namespace QueryMutator.Core
 {
     public class IgnoreMemberBinding : MemberBindingBase
     {
-        public override Expression GenerateExpression(ParameterExpression parameter)
+        public override Expression GenerateExpression()
         {
             return null;
         }
