+++ src/QueryMutator.Core/MemberBindings/MemberBindingBase.cs
@@ -12,6 +12,6 @@ namespace QueryMutator.Core
 
         public MemberInfo TargetMember { get; set; }
 
-        public abstract Expression GenerateExpression(ParameterExpression parameter);
+        public abstract Expression GenerateExpression();
     }
 }
