+++ src/QueryMutator.Core/MemberBindings/MemberListBinding.cs
@@ -7,7 +7,7 @@ namespace QueryMutator.Core
     {
         public new MethodCallExpression SourceExpression { get; set; }
 
-        public override Expression GenerateExpression(ParameterExpression parameter)
+        public override Expression GenerateExpression()
         {
             return SourceExpression;
         }
